backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain-openai
langchain-core
langgraph
langsmith
openai
pandas
python-dotenv
requests
pillow
